backend/requirements.txt
Flask==2.0.1                     # Для создания API
requests==2.26.0                 # Для отправки HTTP-запросов
python-dotenv==0.19.0            # Для работы с .env файлом
gspread==5.1.1                   # Для работы с Google Sheets
oauth2client==4.1.3              # Для авторизации в Google Sheets
mistralai                        # Клиент Mistral AI (установите командой ниже)
python-telegram-bot==13.7        # Для взаимодействия с Telegram API
Werkzeug==2.0.3